refactor: removed unnecessary bot abstraction layers and implemented restart
- moved send message to main.py
- removed bots.py
- added restart password to env
- added curl to dockerfile, for auto restart curling inside container
- added minimum length of messages
- removed all the restart utils and its implementations
