+++ packages/telegram_bots/config.py
@@ -10,3 +10,4 @@ NLP_BOT_TOKEN = os.getenv("NLP_BOT_TOKEN")
 TELEGRAM_BOTS_URL = os.getenv("TELEGRAM_BOTS_URL")
 BACKEND_URL = os.getenv("BACKEND_URL")
 REPORT_PASSWORD = os.getenv("REPORT_PASSWORD")
+RESTART_PASSWORD = os.getenv("RESTART_PASSWORD")
